create splash ui and format packages
