+++ app/src/main/java/com/cris/nvh/moviedb/MainActivity.java
@@ -1,5 +1,7 @@
 package com.cris.nvh.moviedb;
 
+import android.content.Context;
+import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
@@ -9,5 +11,9 @@ public class MainActivity extends AppCompatActivity {
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
 		setContentView(R.layout.activity_main);
+	}
+
+	public static Intent getMainActivityIntent(Context context){
+		return new Intent(context, MainActivity.class);
 	}
 }
