+++ app/src/main/java/com/cris/nvh/moviedb/data/model/Genre.java
@@ -1,4 +1,4 @@
-package com.cris.nvh.moviedb.data.model.data;
+package com.cris.nvh.moviedb.data.model;
 
 import android.os.Parcel;
 import android.os.Parcelable;
@@ -10,28 +10,24 @@ import com.google.gson.annotations.SerializedName;
  * Contact: toiyeuthethao1997@gmail.com
  */
 
-public class Video implements Parcelable {
+public class Genre implements Parcelable {
 	@SerializedName("id")
 	private int mId;
-	@SerializedName("key")
-	private String mKey;
 	@SerializedName("name")
 	private String mName;
 
-	public Video(int id) {
+	public Genre(int id){
 		mId = id;
 	}
 
-	protected Video(Parcel in) {
+	protected Genre(Parcel in) {
 		mId = in.readInt();
-		mKey = in.readString();
 		mName = in.readString();
 	}
 
 	@Override
 	public void writeToParcel(Parcel parcel, int i) {
 		parcel.writeInt(mId);
-		parcel.writeString(mKey);
 		parcel.writeString(mName);
 	}
 
@@ -40,24 +36,20 @@ public class Video implements Parcelable {
 		return 0;
 	}
 
-	public static final Creator<Video> CREATOR = new Creator<Video>() {
+	public static final Creator<Genre> CREATOR = new Creator<Genre>() {
 		@Override
-		public Video createFromParcel(Parcel in) {
-			return new Video(in);
+		public Genre createFromParcel(Parcel in) {
+			return new Genre(in);
 		}
 
 		@Override
-		public Video[] newArray(int size) {
-			return new Video[size];
+		public Genre[] newArray(int size) {
+			return new Genre[size];
 		}
 	};
 
 	public int getId() {
 		return mId;
-	}
-
-	public String getKey() {
-		return mKey;
 	}
 
 	public String getName() {
