+++ app/src/main/java/com/cris/nvh/moviedb/util/Constant.java
@@ -1,0 +1,10 @@
+package com.cris.nvh.moviedb.util;
+
+/**
+ * Created by nvh
+ * Contact: toiyeuthethao1997@gmail.com
+ */
+
+public final class Constant {
+	public static final String BASE_URL = "https://api.themoviedb.org/3/";
+}
